sentinet_py/requirements.txt
numpy
pyzmq
